requirements.txt
# Core AI and Email Processing
beautifulsoup4>=4.11.0
python-dateutil>=2.8.0
python-dotenv>=0.19.0
faiss-cpu>=1.7.0
numpy>=1.21.0
pyttsx3>=2.90
SpeechRecognition>=3.8.0
tzlocal>=4.0
groq>=0.4.0
langchain>=0.0.200
chime>=0.7.0

# Google APIs
google-auth-oauthlib>=0.4.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth>=2.0.0

# Web Framework for Backend API
Flask>=2.0.0
Flask-CORS>=3.0.0

# Additional System Dependencies  
psutil>=5.8.0
requests>=2.25.0
certifi>=2021.5.25

# Optional UI Framework (for legacy PyQt6 interface)
# Uncomment the lines below if you want to use the PyQt6 interface
# PyQt6>=6.4.0
# qt-material>=2.14
